q360_backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0